requirements.txt
asgiref==3.3.0
Django==3.1.3
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
asgiref==3.3.0
click==7.1.2
Django==3.1.3
djangorestframework==3.12.2
ffpb==0.3.0
future==0.18.2
gunicorn==20.0.4
h11==0.11.0
httptools==0.1.1
python-dotenv==0.15.0
pytz==2020.4
sh==1.14.1
sqlparse==0.4.1
tqdm==4.51.0
uvicorn==0.12.2
uvloop==0.14.0

black==20.8b1
pylint==2.6.0
pylint-django==2.3.0
